Add new validation helper file
- Move existing validation helpers to new file

[#11]
